extract out `eslint-config-prettier` logic
