+++ adders/prettier/config/adder.ts
@@ -1,5 +1,6 @@
 import { dedent, defineAdderConfig, log, colors } from "@svelte-add/core";
 import { options } from "./options.js";
+import { addEslintConfigPrettier } from "../../common.js";
 
 export const adder = defineAdderConfig({
     metadata: {
@@ -97,43 +98,7 @@ export const adder = defineAdderConfig({
                 }
                 return hasEslint(deps);
             },
-            content: ({ ast, imports, exports, common }) => {
-                // TODO: maybe this could be more intelligent and we can detect the name of the default import?
-                imports.addDefault(ast, "eslint-plugin-svelte", "svelte");
-                imports.addDefault(ast, "eslint-config-prettier", "prettier");
-
-                const fallbackConfig = common.expressionFromString("[]");
-                const defaultExport = exports.defaultExport(ast, fallbackConfig);
-                const array = defaultExport.value;
-                if (array.type !== "ArrayExpression") return;
-
-                const prettier = common.expressionFromString("prettier");
-                const sveltePrettierConfig = common.expressionFromString("svelte.configs['flat/prettier']");
-                const configSpread = common.createSpreadElement(sveltePrettierConfig);
-
-                const nodesToInsert = [];
-                if (!common.hasNode(array, prettier)) nodesToInsert.push(prettier);
-                if (!common.hasNode(array, configSpread)) nodesToInsert.push(configSpread);
-
-                // finds index of `...svelte.configs["..."]`
-                const idx = array.elements.findIndex(
-                    (el) =>
-                        el?.type === "SpreadElement" &&
-                        el.argument.type === "MemberExpression" &&
-                        el.argument.object.type === "MemberExpression" &&
-                        el.argument.object.property.type === "Identifier" &&
-                        el.argument.object.property.name === "configs" &&
-                        el.argument.object.object.type === "Identifier" &&
-                        el.argument.object.object.name === "svelte",
-                );
-
-                if (idx !== -1) {
-                    array.elements.splice(idx + 1, 0, ...nodesToInsert);
-                } else {
-                    // append to the end as a fallback
-                    array.elements.push(...nodesToInsert);
-                }
-            },
+            content: addEslintConfigPrettier,
         },
     ],
 });
